EPI-73: Other admin panes (#2617)
