+++ packages/desktop/app/components/Sidebar/config.js
@@ -192,22 +192,9 @@ export const FACILITY_MENU_ITEMS = [
 
 export const SYNC_MENU_ITEMS = [
   {
-    key: 'settings',
-    label: 'Settings',
-    path: '/admin/settings',
-    ability: { subject: 'user', action: 'read' },
-  },
-  {
-    key: 'users',
-    label: 'Users',
-    path: '/admin/users',
-    ability: { action: 'read', subject: 'user' },
-  },
-  {
-    key: 'locations',
-    label: 'Locations',
-    path: '/admin/locations',
-    ability: { action: 'read', subject: 'location' },
+    key: 'refdata',
+    label: 'Data Import',
+    path: '/admin/refdata',
   },
   {
     key: 'permissions',
@@ -219,10 +206,5 @@ export const SYNC_MENU_ITEMS = [
     key: 'programs',
     label: 'Programs',
     path: '/admin/programs',
-  },
-  {
-    key: 'refdata',
-    label: 'Data Import',
-    path: '/admin/refdata',
   },
 ];
