+++ packages/desktop/app/views/administration/PermissionsAdminView.js
@@ -1,24 +1,6 @@
-import React, { memo, useCallback } from 'react';
-import styled from 'styled-components';
+import React from 'react';
+import { ImporterView } from './components/ImporterView';
 
-import { useApi } from '../../api';
-import { DataDocumentUploadForm } from './DataDocumentUploadForm';
-
-const Container = styled.div`
-  padding: 32px;
-`;
-
-export const PermissionsAdminView = memo(() => {
-  const api = useApi();
-  const onSubmit = useCallback(
-    ({ file, ...data }) => api.postWithFileUpload('admin/importPermissions', file, data),
-    [api],
-  );
-
-  return (
-    <Container>
-      <h1>Permissions admin</h1>
-      <DataDocumentUploadForm onSubmit={onSubmit} />
-    </Container>
-  );
-});
+export const PermissionsAdminView = () => (
+  <ImporterView title="Import permissions" endpoint="admin/importRefData" />
+);
