+++ packages/desktop/app/views/administration/ProgramsAdminView.js
@@ -1,20 +1,6 @@
-import React, { memo, useCallback } from 'react';
-import { ContentPane } from '../../components';
+import React from 'react';
+import { ImporterView } from './components/ImporterView';
 
-import { useApi } from '../../api';
-import { DataDocumentUploadForm } from './DataDocumentUploadForm';
-
-export const ProgramsAdminView = memo(() => {
-  const api = useApi();
-  const onSubmit = useCallback(
-    ({ file, ...data }) => api.postWithFileUpload('admin/importProgram', file, data),
-    [api],
-  );
-
-  return (
-    <ContentPane>
-      <h1>Programs admin</h1>
-      <DataDocumentUploadForm onSubmit={onSubmit} />
-    </ContentPane>
-  );
-});
+export const ProgramsAdminView = () => (
+  <ImporterView title="Import programs/surveys" endpoint="admin/importProgram" />
+);
