+++ packages/sync-server/app/admin/index.js
@@ -2,10 +2,6 @@ import express from 'express';
 
 import { createDataImporterEndpoint } from './importerEndpoint';
 
-import {
-  importer as permissionsImporter,
-  PERMISSIONS as PERMISSIONS_PERMISSIONS,
-} from './permissionsImporter';
 import { importer as programImporter, PERMISSIONS as PROGRAM_PERMISSIONS } from './programImporter';
 import { importer as refdataImporter, PERMISSIONS as REFDATA_PERMISSIONS } from './refdataImporter';
 
@@ -14,11 +10,6 @@ export const adminRoutes = express.Router();
 adminRoutes.post(
   '/importRefData',
   createDataImporterEndpoint(refdataImporter, REFDATA_PERMISSIONS),
-);
-
-adminRoutes.post(
-  '/importPermissions',
-  createDataImporterEndpoint(permissionsImporter, PERMISSIONS_PERMISSIONS),
 );
 
 adminRoutes.post(
