+++ packages/sync-server/app/admin/programImporter/idify.js
@@ -1,0 +1,1 @@
+export const idify = name => name.toLowerCase().replace(/\W/g, '');
