+++ packages/sync-server/app/admin/refdataImporter/index.js
@@ -1,32 +1,18 @@
-import { singularize } from 'inflection';
-import { camelCase, upperFirst, lowerCase } from 'lodash';
-import { Sequelize } from 'sequelize';
+import { upperFirst } from 'lodash';
 import { readFile } from 'xlsx';
 
 import { log } from 'shared/services/logging';
 import { REFERENCE_TYPE_VALUES } from 'shared/constants';
 
-import {
-  referenceDataLoaderFactory,
-  loaderFactory,
-} from './loaders';
-import { DryRun } from './errors';
+import { normaliseSheetName } from '../importerEndpoint';
+
+import { referenceDataLoaderFactory, loaderFactory } from './loaders';
 import { importSheet } from './sheet';
-import { coalesceStats } from './stats';
 import DEPENDENCIES from './dependencies';
 
-export const PERMISSIONS = ['User', 'ReferenceData'];
+export const PERMISSIONS = ['Permission', 'Role', 'User', 'ReferenceData'];
 
-function normalise(name) {
-  const norm = singularize(camelCase(singularize(lowerCase(name))));
-
-  if (norm === 'placesOfBirth') return 'placeOfBirth';
-  if (norm === 'vaccineSchedule') return 'scheduledVaccine';
-
-  return norm;
-}
-
-async function importerInner({ errors, models, stats, file, whitelist = [] }) {
+export async function importer({ errors, models, stats, file, whitelist = [] }) {
   log.info('Importing data definitions from file', { file });
 
   log.debug('Parse XLSX workbook');
@@ -35,7 +21,7 @@ async function importerInner({ errors, models, stats, file, whitelist = [] }) {
   log.debug('Normalise all sheet names for lookup');
   const sheets = new Map();
   for (const [sheetName, sheet] of Object.entries(workbook.Sheets)) {
-    const name = normalise(sheetName);
+    const name = normaliseSheetName(sheetName);
 
     if (whitelist.length && !whitelist.includes(name)) {
       log.debug('Sheet has been manually excluded', { name });
@@ -91,7 +77,7 @@ async function importerInner({ errors, models, stats, file, whitelist = [] }) {
 
   // sort by length of needs, so that stuff that doesn't depend on anything else gets done first
   // (as an optimisation, the algorithm doesn't need this, but it saves a few cycles)
-  const dataTypes = Object.entries(DEPENDENCIES).map(([k, v]) => [normalise(k), v]);
+  const dataTypes = Object.entries(DEPENDENCIES).map(([k, v]) => [normaliseSheetName(k), v]);
   // eslint-disable-next-line no-unused-vars
   dataTypes.sort(([_ka, a], [_kb, b]) => (a.needs?.length ?? 0) - (b.needs?.length ?? 0));
 
@@ -136,61 +122,8 @@ async function importerInner({ errors, models, stats, file, whitelist = [] }) {
     );
     importedData.push(dataType);
   }
-  
+
   if (!loopProtection) throw new Error('Loop, cycle, or unresolvable import dependencies');
 
   log.debug('Done importing data', { importedData, droppedData });
-}
-
-export async function importer({
-  models,
-  file,
-  dryRun = false,
-  whitelist = [],
-}) {
-  const errors = [];
-  const stats = [];
-
-  try {
-    log.debug('Starting transaction');
-    await models.ReferenceData.sequelize.transaction(
-      {
-        // strongest level to be sure to read/write good data
-        isolationLevel: Sequelize.Transaction.ISOLATION_LEVELS.SERIALIZABLE,
-      },
-      async () => {
-        await importerInner({ errors, models, stats, file, whitelist });
-        if (errors.length > 0) throw new Error('rollback on errors');
-        if (dryRun) throw new DryRun(); // roll back the transaction
-      },
-    );
-    log.debug('Ended transaction');
-
-    if (dryRun) {
-      throw new Error('Data import completed but it was a dry run!!!');
-    } else {
-      return { errors: [], stats: coalesceStats(stats) };
-    }
-  } catch (err) {
-    log.error(`while importing refdata: ${err.stack}`);
-    if (dryRun && err instanceof DryRun) {
-      return {
-        didntSendReason: 'dryRun',
-        errors: [],
-        stats: coalesceStats(stats),
-      };
-    }
-    if (errors.length) {
-      return {
-        didntSendReason: 'validationFailed',
-        errors,
-        stats: coalesceStats(stats),
-      };
-    }
-    return {
-      didntSendReason: 'validationFailed',
-      errors: [err],
-      stats: coalesceStats(stats),
-    };
-  }
 }
