+++ packages/sync-server/app/admin/refdataImporter/sheet.js
@@ -1,17 +1,8 @@
-import { camelCase, lowerCase, lowerFirst, startCase, upperFirst } from 'lodash';
-import { Op } from 'sequelize';
 import { utils } from 'xlsx';
-import { ValidationError as YupValidationError } from 'yup';
 
-import {
-  DataLoaderError,
-  ForeignkeyResolutionError,
-  UpsertionError,
-  ValidationError,
-  WorkSheetError,
-} from './errors';
-import * as schemas from './schemas';
-import { newStatsRow } from './stats';
+import { DataLoaderError, ValidationError, WorkSheetError } from '../errors';
+import { statkey, updateStat } from '../stats';
+import { importRows } from '../importRows';
 
 const FOREIGN_KEY_SCHEMATA = {
   CertifiableVaccine: [
@@ -67,24 +58,8 @@ const FOREIGN_KEY_SCHEMATA = {
   ],
 };
 
-function findFieldName(values, fkField) {
-  const fkFieldLower = fkField.toLowerCase();
-  const fkFieldCamel = camelCase(fkField);
-  const fkFieldUcfirst = upperFirst(fkField);
-  const fkFieldSplit = lowerCase(fkField);
-  const fkFieldSplitUcwords = startCase(fkFieldSplit);
-  if (values[fkField]) return fkField;
-  if (values[fkFieldLower]) return fkFieldLower;
-  if (values[fkFieldCamel]) return fkFieldCamel;
-  if (values[fkFieldUcfirst]) return fkFieldUcfirst;
-  if (values[fkFieldSplit]) return fkFieldSplit;
-  if (values[fkFieldSplitUcwords]) return fkFieldSplitUcwords;
-  return null;
-}
-
 export async function importSheet({ errors, log, models }, { loader, sheetName, sheet }) {
   const stats = {};
-  const statkey = model => (model === 'ReferenceData' ? `${model}/${sheetName}` : model);
 
   log.debug('Loading rows from sheet');
   let sheetRows;
@@ -115,7 +90,7 @@ export async function importSheet({ errors, log, models }, { loader, sheetName, 
           idCache.add(values.id);
         }
 
-        stats[statkey(model)] = stats[statkey(model)] || newStatsRow();
+        updateStat(stats, statkey(model, sheetName), 'created', 0);
         tableRows.push({ model, sheetRow, values });
       }
     } catch (err) {
@@ -123,151 +98,10 @@ export async function importSheet({ errors, log, models }, { loader, sheetName, 
     }
   }
 
-  if (tableRows.length === 0) {
-    log.debug('Nothing left, skipping');
-    return stats;
-  }
-
-  log.debug('Building reverse lookup table');
-  const lookup = new Map();
-  for (const {
-    model,
-    values: { id, type = null, name = null },
-  } of tableRows) {
-    if (!id) continue;
-    const kind = model === 'ReferenceData' ? type : model;
-    lookup.set({ kind, id }, null);
-    if (name) lookup.set({ kind, name: name.toLowerCase() }, id);
-  }
-
-  log.debug('Resolving foreign keys', { rows: tableRows.length });
-  const resolvedRows = [];
-  for (const { model, sheetRow, values } of tableRows) {
-    try {
-      for (const fkSchema of FOREIGN_KEY_SCHEMATA[model] ?? []) {
-        const fkFieldName = findFieldName(values, fkSchema.field);
-        if (fkFieldName) {
-          const fkFieldValue = values[fkFieldName];
-          const fkNameLowerId = `${lowerFirst(fkSchema.field)}Id`;
-
-          const hasLocalId = lookup.has({ kind: fkSchema.field, id: fkFieldValue });
-          const idByLocalName = lookup.get({
-            kind: fkSchema.field,
-            name: fkFieldValue.toLowerCase(),
-          });
-
-          if (hasLocalId) {
-            delete values[fkFieldName];
-            values[fkNameLowerId] = fkFieldValue;
-          } else if (idByLocalName) {
-            delete values[fkFieldName];
-            values[fkNameLowerId] = idByLocalName;
-          } else {
-            const hasRemoteId =
-              (fkSchema.model === 'ReferenceData'
-                ? await models.ReferenceData.count({
-                    where: { type: fkSchema.types, id: fkFieldValue },
-                  })
-                : await models[fkSchema.model].count({ where: { id: fkFieldValue } })) > 0;
-
-            const idByRemoteName = (fkSchema.model === 'ReferenceData'
-              ? await models.ReferenceData.findOne({
-                  where: { type: fkSchema.types, name: { [Op.iLike]: fkFieldValue } },
-                })
-              : await models[fkSchema.model].findOne({
-                  where: {
-                    name: { [Op.iLike]: fkFieldValue },
-                  },
-                })
-            )?.id;
-
-            if (hasRemoteId) {
-              delete values[fkFieldName];
-              values[fkNameLowerId] = fkFieldValue;
-            } else if (idByRemoteName) {
-              delete values[fkFieldName];
-              values[fkNameLowerId] = idByRemoteName;
-            } else {
-              throw new Error(
-                `valid foreign key expected in column ${fkFieldName} (corresponding to ${fkNameLowerId}) but found: ${fkFieldValue}`,
-              );
-            }
-          }
-        }
-      }
-
-      resolvedRows.push({ model, sheetRow, values });
-    } catch (err) {
-      stats[statkey(model)].errored += 1;
-      errors.push(new ForeignkeyResolutionError(sheetName, sheetRow, err));
-    }
-  }
-
-  if (resolvedRows.length === 0) {
-    log.debug('Nothing left, skipping');
-    return stats;
-  }
-
-  log.debug('Validating data', { rows: resolvedRows.length });
-  const validRows = [];
-  for (const { model, sheetRow, values } of resolvedRows) {
-    try {
-      let schemaName;
-      if (model === 'ReferenceData') {
-        const specificSchemaName = `RD${sheetName}`;
-        const specificSchemaExists = !!schemas[specificSchemaName];
-        if (specificSchemaExists) {
-          schemaName = specificSchemaName;
-        } else {
-          schemaName = 'ReferenceData';
-        }
-      } else {
-        const specificSchemaExists = !!schemas[model];
-        if (specificSchemaExists) {
-          schemaName = model;
-        } else {
-          schemaName = 'Base';
-        }
-      }
-
-      const schema = schemas[schemaName];
-      validRows.push({
-        model,
-        sheetRow,
-        values: await schema.validate(values, { abortEarly: false }),
-      });
-    } catch (err) {
-      stats[statkey(model)].errored += 1;
-      if (err instanceof YupValidationError) {
-        for (const valerr of err.errors) {
-          errors.push(new ValidationError(sheetName, sheetRow, valerr));
-        }
-      }
-    }
-  }
-
-  if (validRows.length === 0) {
-    log.debug('Nothing left, skipping');
-    return stats;
-  }
-
-  log.debug('Upserting database rows', { rows: validRows.length });
-  for (const { model, sheetRow, values } of validRows) {
-    const Model = models[model];
-    const existing = values.id && (await Model.findByPk(values.id));
-    try {
-      if (existing) {
-        await existing.update(values);
-        stats[statkey(model)].updated += 1;
-      } else {
-        await Model.create(values);
-        stats[statkey(model)].created += 1;
-      }
-    } catch (err) {
-      stats[statkey(model)].errored += 1;
-      errors.push(new UpsertionError(sheetName, sheetRow, err));
-    }
-  }
+  await importRows(
+    { errors, log, models },
+    { rows: tableRows, sheetName, stats, foreignKeySchemata: FOREIGN_KEY_SCHEMATA },
+  );
 
   return stats;
 }
