+++ packages/sync-server/app/admin/stats.js
@@ -1,8 +1,19 @@
 import { log } from 'shared/services/logging';
 
-export function newStatsRow({ created = 0, updated = 0, errored = 0 } = {}) {
+export function statkey(model, sheetName) {
+  return model === 'ReferenceData' ? `${model}/${sheetName}` : model;
+}
+
+function newStatsRow({ created = 0, updated = 0, errored = 0 } = {}) {
   return { created, updated, errored };
 }
+
+/* eslint-disable no-param-reassign */
+export function updateStat(stats, key, field, incr = 1) {
+  stats[key] = stats[key] || newStatsRow();
+  stats[key][field] += incr;
+}
+/* eslint-enable no-param-reassign */
 
 export function coalesceStats(stats) {
   const allStats = {};
